update-hanayama

+++ app/database_function/review_sql.php
@@ -8,7 +8,7 @@ function registerReview($userId, $shopId, $reviewNum, $review) {
 
 function checkExistsReview($userId, $shopId) {
     $dbh = dbConnection::getConnection();
-    $sql = 'select userid ' . TABLE_NAME_REVIEWS . ' where ? = pgp_sym_decrypt(?, \'' . getenv('DB_ENCRYPT_PASS') . '\') and ? = shopid';
+    $sql = 'select userid ' . TABLE_NAME_REVIEWS . ' where ? = pgp_sym_decrypt(userid, \'' . getenv('DB_ENCRYPT_PASS') . '\') and ? = shopid';
     $sth = $dbh->prepare($sql);
     $sth->execute(array($userId, $shopId));
     // if no record
@@ -29,7 +29,7 @@ function updateReview($userId) {
 
 function deleteReview($userId, $shopId) {
     $dbh = dbConnection::getConnection();
-    $sql = 'delete from '. TABLE_NAME_REVIEWS .' where ? = pgp_sym_decrypt(?, \'' . getenv('DB_ENCRYPT_PASS') . '\') and ? = shopid';
+    $sql = 'delete from '. TABLE_NAME_REVIEWS .' where ? = pgp_sym_decrypt(userid, \'' . getenv('DB_ENCRYPT_PASS') . '\') and ? = shopid';
     $sth = $dbh->prepare($sql);
     $sth->execute(array($userId, $shopId, $reviewNum, $review));
 }
